indexer/requirements.txt
pdfplumber>=0.10.0
openai>=1.0.0
sqlite-vss>=0.1.2
watchdog>=3.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
